+++ mindsdb/integrations/handlers/footballApi_handler/footballapi_constants.py
@@ -1,0 +1,3 @@
+FOOTBALL_API_CLIENT_METHODS = {
+    "PLAYERS": "get_players"
+}
